# openapi.yaml
openapi: 3.0.3
info:
  title: 餐厅微服务 API
  description: OpenAPI 3.0 描述 - 包含用户、菜品、订单服务的主要接口与示例
  version: "1.0.0"
servers:
  - url: http://localhost:8900
    description: 本地开发服务器
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse_Order:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Success
        data:
          $ref: '#/components/schemas/Order'
        timestamp:
          type: string
          format: date-time
    ApiResponse_OrderDetail:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Success
        data:
          $ref: '#/components/schemas/OrderDetailDTO'
        timestamp:
          type: string
          format: date-time
...
    OrderItemDTO:
      type: object
      properties:
        dishId:
          type: integer
        quantity:
          type: integer
      required: [dishId,quantity]
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNo:
          type: string
        userId:
          type: integer
        totalAmount:
          type: number
          format: double
        status:
          type: string
        remark:
          type: string
        address:
          type: string
        createTime:
          type: string
          format: date-time
    OrderItemDetail:
      type: object
      properties:
        dishId:
          type: integer
        dishName:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
        subtotal:
          type: number
          format: double
    OrderDetailDTO:
      type: object
      properties:
        orderNo:
          type: string
        userId:
          type: integer
        address:
          type: string
        remark:
          type: string
        totalAmount:
          type: number
          format: double
        status:
          type: string
        createTime:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDetail'
    OrderRequest:
      type: object
      properties:
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDTO'
        remark:
          type: string
      required: [userId,items]
paths:
...
  /api/orders/admin/list:
    get:
      summary: 管理员查询所有订单
      tags: [Order]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
  /api/orders:
    post:
      summary: 创建订单
      tags: [Order]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Order'
  /api/orders/id/{orderId}:
    get:
      summary: 根据订单 ID 查询
      tags: [Order]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Order'
  /api/orders/admin/{orderId}/finish:
    put:
      summary: 管理员标记订单完成
      tags: [Order]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Order'
  /api/orders/user/list/{userId}:
    get:
      summary: 查询用户的订单列表
      tags: [Order]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
  /api/orders/admin/status/{status}:
    get:
      summary: 按订单状态查询
      tags: [Order]
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
            enum: [CREATED, FINISHED]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
  /api/orders/orderno/{orderNo}:
    get:
      summary: 根据订单号查询订单详情
      tags: [Order]
      parameters:
        - name: orderNo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_OrderDetail'
...